requirements.txt
numpy
matplotlib
scikit-learn
uproot
pandas
mpl_scatter_density
scipy
pyyaml
tensorflow # the biggun'
pydantic
arviz
mplhep
tqdm
numba